changed int to long

+++ src/main/java/com/bae/accountmaker/service/AccountNumberGeneratorImplementation.java
@@ -8,8 +8,8 @@ import org.springframework.stereotype.Service;
 public class AccountNumberGeneratorImplementation implements AccountNumberGenerator {
 
 	@Override
-	public Integer generateNumber() {
-		Integer number = (int) (Math.random() * 100000000);
+	public Long generateNumber() {
+		Long number = (long) (Math.random() * 100000000);
 		number = number * 100;
 		return number;
 	}
